backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-openai
langchain-core
openai
faiss-cpu
numpy
python-dotenv
requests